Move the Guassian/ Affinity matrix calcs out of PIC. Presently in the test suite
